+++ packages/cli/src/zosfiles/delete/zfs/zfs.handler.ts
@@ -18,7 +18,8 @@ import { IZosFilesOptions, Delete, IZosFilesResponse } from "@zowe/zos-files-for
  */
 export default class ZfsHandler extends ZosFilesBaseHandler {
     public async processWithSession(commandParameters: IHandlerParameters, session: AbstractSession): Promise<IZosFilesResponse> {
-        const zosFilesOptions: IZosFilesOptions = {responseTimeout: commandParameters.arguments.responseTimeout};
+        const zosFilesOptions: IZosFilesOptions = {responseTimeout: commandParameters.arguments.responseTimeout,            tsoAccount: commandParameters.arguments.tsoAccount,
+            tsoProcedure: commandParameters.arguments.tsoProcedure};
         return Delete.zfs(session, commandParameters.arguments.fileSystemName, zosFilesOptions);
     }
 }